Lint and test.
