+++ app/src/androidTest/java/tech/ula/model/MigrationTestHelper.kt
@@ -1,114 +1,54 @@
 package tech.ula.model
 
-import android.arch.persistence.db.SupportSQLiteDatabase
-import android.arch.persistence.db.framework.FrameworkSQLiteOpenHelperFactory
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.testing.MigrationTestHelper
-import android.database.sqlite.SQLiteDatabase
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-import kotlinx.coroutines.experimental.runBlocking
-import org.junit.Assert.*
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import tech.ula.blockingObserve
-import tech.ula.model.entities.Session
-import tech.ula.model.repositories.UlaDatabase
-import tech.ula.model.repositories.Migration1To2
-import tech.ula.model.repositories.Migration2To3
-import java.io.IOException
+import android.content.ContentValues
 
-@RunWith(AndroidJUnit4::class)
-class MigrationTest {
-    private val TEST_DB = "migration-test"
-
-    private val migrationHelper = MigrationTestHelper()
-
-    @get:Rule
-    val helper = MigrationTestHelper(InstrumentationRegistry.getInstrumentation(),
-        UlaDatabase::class.java.canonicalName,
-        FrameworkSQLiteOpenHelperFactory())
-
-    @Test
-    @Throws(IOException::class)
-    fun migrate1To2() {
-        val db = helper.createDatabase(TEST_DB, 1)
-
-        insertVersion1Filesystem(1, "firstFs", db)
-        insertVersion1Session(1, "firstSession", db)
-
-        db.close()
-
-        helper.runMigrationsAndValidate(TEST_DB, 2, true, Migration1To2())
-
-        val migratedDb = getMigratedDatabase()
-        val fs = migratedDb.filesystemDao().getFilesystemByName("firstFs")
-        val session = migratedDb.sessionDao().getSessionByName("firstSession")
-
-        assertFalse(fs.isDownloaded)
-        assertFalse(session.isExtracted)
-        assert(session.lastUpdated == 0L)
-        assert(session.bindings == "")
+class MigrationTestHelper {
+    fun getVersion1Filesystem(id: Long, name: String): ContentValues {
+        val filesystemValues = ContentValues()
+        filesystemValues.put("id", id)
+        filesystemValues.put("name", name)
+        filesystemValues.put("distributionType", "dummy")
+        filesystemValues.put("archType", "dummy")
+        filesystemValues.put("defaultUsername", "dummy")
+        filesystemValues.put("defaultPassword", "dummy")
+        filesystemValues.put("location", "dummy")
+        filesystemValues.put("dateCreated", "dummy")
+        filesystemValues.put("realRoot", 0)
+        return filesystemValues
     }
 
-    @Test
-    @Throws(IOException::class)
-    fun migrate2To3() {
-        val db = helper.createDatabase(TEST_DB, 2)
-
-        insertVersion2Filesystem(1, "fs", db)
-        db.close()
-
-        helper.runMigrationsAndValidate(TEST_DB, 3, true, Migration2To3())
-
-        val migratedDb = getMigratedDatabase()
-
-        val fs = migratedDb.filesystemDao().getFilesystemByName("fs")
-        assertFalse(fs.isAppsFilesystem)
-
-        val session1 = Session(1, name = "test", filesystemId = 1, filesystemName = "fs")
-        val session2 = Session(2, name = "test", filesystemId = 1, filesystemName = "fs", isAppsSession = true)
-        runBlocking {
-            migratedDb.sessionDao().insertSession(session1)
-            migratedDb.sessionDao().insertSession(session2)
-        }
-
-        val sessions = migratedDb.sessionDao().getAllSessions().blockingObserve()!!
-        assertEquals(2, sessions.size)
-        assertTrue(sessions.contains(session1))
-        assertTrue(sessions.contains(session2))
-
-        val returnedSession1 = sessions[0]
-        val returnedSession2 = sessions[1]
-        assertEquals(session1, returnedSession1)
-        assertEquals(session2, returnedSession2)
-        assertFalse(returnedSession1.isAppsSession)
-        assertTrue(returnedSession2.isAppsSession)
+    fun getVersion1Session(id: Long, name: String): ContentValues {
+        val sessionValues = ContentValues()
+        sessionValues.put("id", id)
+        sessionValues.put("name", name)
+        sessionValues.put("filesystemId", 1)
+        sessionValues.put("filesystemName", "firstFs")
+        sessionValues.put("active", 0)
+        sessionValues.put("username", "dummy")
+        sessionValues.put("password", "dummy")
+        sessionValues.put("geometry", "dummy")
+        sessionValues.put("serviceType", "dummy")
+        sessionValues.put("clientType", "dummy")
+        sessionValues.put("port", 0)
+        sessionValues.put("pid", 0)
+        sessionValues.put("startupScript", "dummy")
+        sessionValues.put("runAtDeviceStartup", 0)
+        sessionValues.put("initialCommand", "dummy")
+        return sessionValues
     }
 
-    private fun getMigratedDatabase(): UlaDatabase {
-        val db = Room.databaseBuilder(InstrumentationRegistry.getTargetContext(),
-                UlaDatabase::class.java, TEST_DB)
-                .addMigrations(Migration1To2(), Migration2To3())
-                .build()
-
-        helper.closeWhenFinished(db)
-        return db
-    }
-
-    private fun insertVersion1Filesystem(id: Long, name: String, db: SupportSQLiteDatabase) {
-        val filesystem = migrationHelper.getVersion1Filesystem(id, name)
-        db.insert("filesystem", SQLiteDatabase.CONFLICT_REPLACE, filesystem)
-    }
-
-    private fun insertVersion1Session(id: Long, name: String, db: SupportSQLiteDatabase) {
-        val session = migrationHelper.getVersion1Session(id, name)
-        db.insert("session", SQLiteDatabase.CONFLICT_REPLACE, session)
-    }
-
-    private fun insertVersion2Filesystem(id: Long, name: String, db: SupportSQLiteDatabase) {
-        val filesystem = migrationHelper.getVersion2Filesystem(id, name)
-        db.insert("filesystem", SQLiteDatabase.CONFLICT_REPLACE, filesystem)
+    fun getVersion2Filesystem(id: Long, name: String): ContentValues {
+        val filesystemValues = ContentValues()
+        filesystemValues.put("id", id)
+        filesystemValues.put("name", name)
+        filesystemValues.put("distributionType", "dummy")
+        filesystemValues.put("archType", "dummy")
+        filesystemValues.put("defaultUsername", "dummy")
+        filesystemValues.put("defaultPassword", "dummy")
+        filesystemValues.put("location", "dummy")
+        filesystemValues.put("dateCreated", "dummy")
+        filesystemValues.put("realRoot", 0)
+        filesystemValues.put("isDownloaded", 0)
+        return filesystemValues
     }
 }
